Allow alt="" or role="presentation" (#22)
* Allow alt="" or role="presentation"

As we discussed on #6, it is okay to use `alt=""` without
`role="presentation"`. It is possible that we want to enforce no alt
prop if `role="presentation"` but I'll leave that for another time and
possibly a different rule.

* Slightly simplify img-uses-alt error tests

We don't really get much value out of having two ways to check for
errors. Let's just use the one function everywhere.