Update img-has-alt docs for latest recommendation
Generally speaking, developers should choose to avoid ARIA attributes if
the same thing can be achieved without them. In the case of images and
alt text, we should prefer alt="" over role="presentation". This was
discussed in #100 and fixed in #102, but the documentation wasn't fully
updated to reflect this behavior.

Fixes #162.